+++ src/errors/collect_error.rs
@@ -1,4 +1,6 @@
+#[cfg(feature = "alloc")]
 use alloc::boxed::Box;
+
 use core::error::Error;
 use core::fmt::{Debug, Display, Formatter};
 use core::iter::Chain;
@@ -52,8 +54,10 @@ pub struct CollectError<I, C, E> {
     pub error: E,
 
     /// The internal data of a [`CollectError`].
-    #[cfg(not(doc))]
+    #[cfg(all(not(doc), feature = "alloc"))]
     data: Box<CollectErrorData<I, C, E>>,
+    #[cfg(all(not(doc), not(feature = "alloc")))]
+    data: CollectErrorData<I, C, E>,
 }
 
 /// The internal data of a [`CollectError`].
@@ -90,11 +94,42 @@ impl<I, C, E> CollectError<I, C, E> {
     ///
     /// assert_eq!(error.remain, 1..=3);
     /// assert_eq!(error.collected, vec![1, 2]);
+    /// ```
+    #[must_use]
+    #[cfg(feature = "alloc")]
+    pub fn new(remain: I, collected: C, error: E) -> Self {
+        CollectErrorData { remain, collected, error }.pipe(Box::new).pipe(|data| Self { data })
+    }
+
+    /// Creates a new [`CollectError`].
+    ///
+    /// # Arguments
+    ///
+    /// * `remain` - The remaining iterator after the error occurred.
+    /// * `collected` - The values that were collected.
+    /// * `error` - The error that occurred.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use collect_failable::errors::capacity::{CapacityError, CapacityErrorKind};
+    /// # use collect_failable::errors::types::SizeHint;
+    /// # use collect_failable::errors::CollectError;
+    /// let error = CollectError::new(
+    ///     1..=3,
+    ///     vec![1, 2],
+    ///     CapacityError::<i32>::bounds(SizeHint::exact(2), SizeHint::exact(3)),
+    /// );
+    ///
+    /// assert_eq!(error.remain, 1..=3);
+    /// assert_eq!(error.collected, vec![1, 2]);
     /// assert_eq!(error.error.kind, CapacityErrorKind::Bounds { hint: SizeHint::exact(3) });
     /// assert_eq!(error.error.capacity, SizeHint::exact(2));
     /// ```
+    #[must_use]
+    #[cfg(not(feature = "alloc"))]
     pub fn new(remain: I, collected: C, error: E) -> Self {
-        CollectErrorData { remain, collected, error }.pipe(Box::new).pipe(|data| Self { data })
+        CollectErrorData { remain, collected, error }.pipe(|data| Self { data })
     }
 
     /// Consumes the error, returning a `CollectErrorData` containing the [`CollectError::remain`],
@@ -120,14 +155,63 @@ impl<I, C, E> CollectError<I, C, E> {
     /// assert_eq!(data.error.capacity, SizeHint::exact(2));
     /// ```
     #[must_use]
+    #[cfg(feature = "alloc")]
     pub fn into_data(self) -> CollectErrorData<I, C, E> {
         *self.data
+    }
+
+    /// Consumes the error, returning a `CollectErrorData` containing the [`CollectError::remain`],
+    /// [`CollectError::collected`] values, and [`CollectError::error`].
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use collect_failable::errors::capacity::{CapacityError, CapacityErrorKind};
+    /// # use collect_failable::errors::types::SizeHint;
+    /// # use collect_failable::errors::CollectError;
+    /// let error = CollectError::new(
+    ///     1..=3,
+    ///     vec![1, 2],
+    ///     CapacityError::<i32>::bounds(SizeHint::exact(2), SizeHint::exact(3)),
+    /// );
+    ///
+    /// let data = error.into_data();
+    ///
+    /// assert_eq!(data.remain, 1..=3);
+    /// assert_eq!(data.collected, vec![1, 2]);
+    /// assert_eq!(data.error.kind, CapacityErrorKind::Bounds { hint: SizeHint::exact(3) });
+    /// assert_eq!(data.error.capacity, SizeHint::exact(2));
+    /// ```
+    #[must_use]
+    #[cfg(not(feature = "alloc"))]
+    pub fn into_data(self) -> CollectErrorData<I, C, E> {
+        self.data
     }
 }
 
 /// Consumes the error and creates a new [Iterator] with the data it was
 /// created from: the overflow item (if any), the [`collected`](CollectError::collected),
-/// values and the remaining [`iter`](CollectError::iter), in that order.
+/// values and the remaining [`remain`](CollectError::remain), in that order.
+///
+/// # Examples
+///
+/// ```rust
+/// # use collect_failable::errors::capacity::{CapacityError, CapacityErrorKind};
+/// # use collect_failable::errors::types::SizeHint;
+/// # use collect_failable::errors::CollectError;
+/// let error = CollectError::new(
+///     3..,
+///     vec![1, 2],
+///     CapacityError::<i32>::bounds(SizeHint::exact(2), SizeHint::unbounded(usize::MAX - 3)),
+/// );
+///
+/// let mut iter = error.into_iter();
+///
+/// assert_eq!(iter.next(), Some(1), "should contain first collected item");
+/// assert_eq!(iter.next(), Some(2), "should contain second collected item");
+/// assert_eq!(iter.next(), Some(3), "should contain first remaining item");
+/// // iterator continues indefinitely
+/// ```
 impl<I: Iterator, C, E> IntoIterator for CollectError<I, C, E>
 where
     C: IntoIterator<Item = I::Item>,
@@ -183,6 +267,7 @@ impl<I, C, E: Debug> Debug for CollectErrorData<I, C, E> {
             .finish()
     }
 }
+
 #[doc(hidden)]
 impl<I, C, E: Display> Display for CollectErrorData<I, C, E> {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {