+++ src/errors/mod.rs
@@ -8,37 +8,33 @@ pub mod collision;
 
 mod error_item_provider;
 
-#[cfg(feature = "alloc")]
 mod collect_error;
-#[cfg(feature = "alloc")]
 mod extend_error;
-#[cfg(feature = "alloc")]
 mod result_collection_error;
 
-#[cfg(feature = "tuple")]
+#[cfg(feature = "tuples")]
 mod tuple_extend_error;
-#[cfg(feature = "tuple")]
+#[cfg(feature = "tuples")]
 mod unzip_error;
 
-pub use collision::*;
 pub use error_item_provider::*;
 
-#[cfg(feature = "alloc")]
 pub use collect_error::*;
-#[cfg(feature = "alloc")]
 pub use extend_error::*;
-#[cfg(feature = "alloc")]
 pub use result_collection_error::*;
 
-#[cfg(feature = "tuple")]
+pub use capacity::CapacityError;
+pub use collision::Collision;
+
+#[cfg(feature = "tuples")]
 pub use tuple_extend_error::*;
-#[cfg(feature = "tuple")]
+#[cfg(feature = "tuples")]
 pub use unzip_error::*;
 
 /// Foreign types used by the error types.
 pub mod types {
     /// Re-export of the `Either` type from the `either` crate.
-    #[cfg(feature = "tuple")]
+    #[cfg(feature = "tuples")]
     pub use either::Either;
 
     /// Re-export of the [`SizeHint`] type from the `size_hinter` crate.
@@ -48,5 +44,7 @@ pub mod types {
 /// Types related to [`PartialArray`](crate::impls::unsafe::PartialArray).
 #[cfg(feature = "unsafe")]
 pub mod partial_array {
+    #[doc(hidden)]
+    pub use crate::impls::r#unsafe::{ArrayIndex, PostInc};
     pub use crate::impls::r#unsafe::{Drain, IntoArrayError, PartialArray};
 }