+++ src/errors/unzip_error.rs
@@ -1,8 +1,9 @@
+#[cfg(feature = "alloc")]
+use alloc::boxed::Box;
+
 use core::error::Error;
 use core::fmt::{Debug, Display, Formatter};
 use core::ops::Deref;
-
-use alloc::boxed::Box;
 
 use display_as_debug::fmt::DebugStructExt;
 use display_as_debug::types::{Full, Short};
@@ -46,39 +47,57 @@ pub struct UnzipError<E, Failed, Partial, Pending, I> {
     /// The remaining iterator after the error occurred
     pub remaining: I,
 
-    #[cfg(not(doc))]
-    data: [Box<UnzipErrorData<E, Failed, Partial, Pending, I>>; {
-        /// The internal data of an [`UnzipError`].
-        #[doc(hidden)]
-        pub struct UnzipErrorData<E, Failed, Partial, Pending, I> {
-            /// The error that occurred during extension.
-            pub error: E,
-            /// The partial collection from the failed side.
-            pub failed: Failed,
-            /// The incomplete collection from the successful side.
-            pub partial: Partial,
-            /// The pending item from the successful side, if any.
-            pub pending: Option<Pending>,
-            /// The remaining iterator after the error occurred
-            pub remaining: I,
-        }
-    }],
+    #[cfg(all(not(doc), feature = "alloc"))]
+    data: Box<UnzipErrorData<E, Failed, Partial, Pending, I>>,
+    #[cfg(all(not(doc), not(feature = "alloc")))]
+    data: UnzipErrorData<E, Failed, Partial, Pending, I>,
+}
+
+/// The internal data of an [`UnzipError`].
+#[doc(hidden)]
+pub struct UnzipErrorData<E, Failed, Partial, Pending, I> {
+    /// The error that occurred during extension.
+    pub error: E,
+    /// The partial collection from the failed side.
+    pub failed: Failed,
+    /// The incomplete collection from the successful side.
+    pub partial: Partial,
+    /// The pending item from the successful side, if any.
+    pub pending: Option<Pending>,
+    /// The remaining iterator after the error occurred
+    pub remaining: I,
+}
+
+#[doc(hidden)]
+impl<Failed: TryExtendOne, Partial: TryExtendOne, I: Iterator> UnzipError<Failed::Error, Failed, Partial, Partial::Item, I> {
+    /// Creates a new [`UnzipError`].
+    #[must_use]
+    #[cfg(feature = "alloc")]
+    pub fn new(error: Failed::Error, failed: Failed, partial: Partial, pending: Option<Partial::Item>, remaining: I) -> Self {
+        UnzipErrorData { error, failed, partial, pending, remaining }.pipe(Box::new).pipe(|data| Self { data })
+    }
+
+    /// Creates a new [`UnzipError`].
+    #[must_use]
+    #[cfg(not(feature = "alloc"))]
+    pub fn new(error: Failed::Error, failed: Failed, partial: Partial, pending: Option<Partial::Item>, remaining: I) -> Self {
+        UnzipErrorData { error, failed, partial, pending, remaining }.pipe(|data| Self { data })
+    }
 }
 
 impl<E, Failed, Partial, Pending, I> UnzipError<E, Failed, Partial, Pending, I> {
     /// Consumes the error, returning the data.
     #[must_use]
+    #[cfg(feature = "alloc")]
     pub fn into_data(self) -> UnzipErrorData<E, Failed, Partial, Pending, I> {
         *self.data
     }
-}
 
-#[doc(hidden)]
-impl<Failed: TryExtendOne, Partial: TryExtendOne, I: Iterator> UnzipError<Failed::Error, Failed, Partial, Partial::Item, I> {
-    #[doc(hidden)]
-    /// Creates a new [`UnzipError`].
-    pub fn new(error: Failed::Error, failed: Failed, partial: Partial, pending: Option<Partial::Item>, remaining: I) -> Self {
-        UnzipErrorData { error, failed, partial, pending, remaining }.pipe(Box::new).pipe(|data| Self { data })
+    /// Consumes the error, returning the data.
+    #[must_use]
+    #[cfg(not(feature = "alloc"))]
+    pub fn into_data(self) -> UnzipErrorData<E, Failed, Partial, Pending, I> {
+        self.data
     }
 }