+++ src/impls/unsafe/mod.rs
@@ -1,4 +1,7 @@
 mod array;
+mod array_index;
 mod partial_array;
 
+#[doc(hidden)]
+pub use array_index::{ArrayIndex, PostInc};
 pub use partial_array::{Drain, IntoArrayError, PartialArray};