+++ src/traits/mod.rs
@@ -1,11 +1,11 @@
 mod try_extend;
 mod try_from_iterator;
 
-#[cfg(feature = "tuple")]
+#[cfg(feature = "tuples")]
 mod try_unzip;
 
 pub use try_extend::*;
 pub use try_from_iterator::*;
 
-#[cfg(feature = "tuple")]
+#[cfg(feature = "tuples")]
 pub use try_unzip::TryUnzip;