+++ tests/impl/try_unzip.rs
@@ -1,11 +1,11 @@
+use collect_failable::TryUnzip;
+use std::collections::HashSet;
+
 const INVALID_DATA_A: [(u32, u32); 2] = [(1, 3), (1, 4)];
 const INVALID_DATA_B: [(u32, u32); 2] = [(1, 2), (2, 2)];
 
 #[test]
 fn try_unzip_success() {
-    use collect_failable::TryUnzip;
-    use std::collections::HashSet;
-
     let data = vec![(1, 10), (2, 20), (3, 30)];
     let (set_a, set_b) = data.into_iter().try_unzip::<HashSet<_>, HashSet<_>>().expect("Should succeed");
 
@@ -15,9 +15,6 @@ fn try_unzip_success() {
 
 #[test]
 fn try_unzip_fail_a() {
-    use collect_failable::TryUnzip;
-    use std::collections::HashSet;
-
     let err = INVALID_DATA_A.into_iter().try_unzip::<HashSet<_>, HashSet<_>>().expect_err("Should be Err");
     let err = err.left().expect("Should be left");
 
@@ -26,9 +23,6 @@ fn try_unzip_fail_a() {
 
 #[test]
 fn try_unzip_fail_b() {
-    use collect_failable::TryUnzip;
-    use std::collections::HashSet;
-
     let err = INVALID_DATA_B.into_iter().try_unzip::<HashSet<_>, HashSet<_>>().expect_err("Should be Err");
     let err = err.right().expect("Should be right");
 
@@ -37,9 +31,6 @@ fn try_unzip_fail_b() {
 
 #[test]
 fn try_unzip_recover_partial_b_on_a_failure() {
-    use collect_failable::TryUnzip;
-    use std::collections::HashSet;
-
     // When FromA fails, the incomplete FromB should be recoverable
     let data = vec![(1, 10), (2, 20), (1, 30), (3, 40)]; // Collision on first element (A)
     let err = data.into_iter().try_unzip::<HashSet<_>, HashSet<_>>().expect_err("Should fail");
@@ -53,9 +44,6 @@ fn try_unzip_recover_partial_b_on_a_failure() {
 
 #[test]
 fn try_unzip_recover_partial_a_on_b_failure() {
-    use collect_failable::TryUnzip;
-    use std::collections::HashSet;
-
     // When FromB fails, the incomplete FromA should be recoverable
     let data = vec![(10, 1), (20, 2), (30, 2), (40, 3)]; // Collision on second element (B)
     let err = data.into_iter().try_unzip::<HashSet<_>, HashSet<_>>().expect_err("Should fail");